FEATURE: Update to latest algolia gem (#33)
This commit includes several bug fixes and improvements. The indexing
process is more efficient and uses less Algolia resources.

Security-wise, the admin API key is no longer visible to the client.
Records that are no longer visible will also be deleted from the index.
It used to use system user's guardian which meant that everything was
indexed. The new default is to use the guardian of the anonymous user.

The code was reorganized and logic for each model was moved in their
own class. Tests were added for indexers.

Co-authored-by: Dan Ungureanu <dan@ungureanu.me>